+++ packages/hydrogen/src/pagination/Pagination.ts
@@ -1,10 +1,16 @@
-import {createElement, useEffect, useMemo, useState} from 'react';
+import {createElement, useEffect, useMemo, useRef, useState} from 'react';
 import type {
   Maybe,
   PageInfo,
 } from '@shopify/hydrogen-react/storefront-api-types';
 import {flattenConnection} from '@shopify/hydrogen-react';
-import {Link, LinkProps, useNavigation, useLocation} from '@remix-run/react';
+import {
+  Link,
+  LinkProps,
+  useNavigation,
+  useLocation,
+  useNavigate,
+} from '@remix-run/react';
 
 type Connection<NodesType> =
   | {
@@ -146,6 +152,23 @@ export function Pagination<NodesType>({
   });
 }
 
+function getParamsWithoutPagination(paramsString?: string) {
+  const params = new URLSearchParams(paramsString);
+  params.delete('cursor');
+  params.delete('direction');
+  return params.toString();
+}
+
+function makeError(prop: string) {
+  throw new Error(
+    `The Pagination component requires ${
+      '`' + prop + '`'
+    } to be a part of your query. See the guide on how to setup your query to include ${
+      '`' + prop + '`'
+    }: https://shopify.dev/docs/custom-storefronts/hydrogen/data-fetching/pagination#setup-the-paginated-query`,
+  );
+}
+
 /**
  * Get cumulative pagination logic for a given connection
  */
@@ -158,9 +181,31 @@ export function usePagination<NodesType>(
   startCursor: Maybe<string> | undefined;
   endCursor: Maybe<string> | undefined;
 } {
-  const {state, search} = useLocation() as {
+  if (!connection.pageInfo) {
+    makeError('pageInfo');
+  }
+
+  if (typeof connection.pageInfo.startCursor === 'undefined') {
+    makeError('pageInfo.startCursor');
+  }
+
+  if (typeof connection.pageInfo.endCursor === 'undefined') {
+    makeError('pageInfo.endCursor');
+  }
+
+  if (typeof connection.pageInfo.hasNextPage === 'undefined') {
+    makeError('pageInfo.hasNextPage');
+  }
+
+  if (typeof connection.pageInfo.hasPreviousPage === 'undefined') {
+    makeError('pageInfo.hasPreviousPage');
+  }
+
+  const navigate = useNavigate();
+  const {state, search, pathname} = useLocation() as {
     state?: PaginationState<NodesType>;
     search?: string;
+    pathname?: string;
   };
 
   const params = new URLSearchParams(search);
@@ -174,41 +219,60 @@ export function usePagination<NodesType>(
     hasNextPage: connection.pageInfo.hasNextPage,
   });
 
+  // Keep track of the current URL state, to compare whenever the URL changes
+  const urlRef = useRef({
+    params: getParamsWithoutPagination(search),
+    pathname,
+  });
+
   // Within an effect to prevent hydration errors
   useEffect(() => {
-    if (state?.nodes) {
+    if (
+      // If the URL changes (independent of pagination params)
+      // then reset the pagination params in the URL
+      getParamsWithoutPagination(search) !== urlRef.current.params ||
+      pathname !== urlRef.current.pathname
+    ) {
+      urlRef.current = {
+        pathname,
+        params: getParamsWithoutPagination(search),
+      };
+      navigate(`${pathname}?${getParamsWithoutPagination(search)}`, {
+        replace: true,
+        preventScrollReset: true,
+        state: {nodes: undefined, pageInfo: undefined},
+      });
+    } else if (state?.nodes) {
       // Take existing nodes from history.state and append or prepend new nodes from the connection
       setNodes(
         isPrevious
           ? [...flattenConnection(connection), ...state.nodes]
           : [...state.nodes, ...flattenConnection(connection)],
       );
-    }
 
-    if (state?.pageInfo) {
-      // Based on the direction we are going, only take the connection pageInfo in that direction,
-      // otherwise keep the existing pageInfo from history.state
-      let pageStartCursor =
-        state?.pageInfo?.startCursor === undefined
-          ? connection.pageInfo.startCursor
-          : state.pageInfo.startCursor;
+      if (state?.pageInfo) {
+        // Based on the direction we are going, only take the connection pageInfo in that direction,
+        // otherwise keep the existing pageInfo from history.state
+        let pageStartCursor =
+          state?.pageInfo?.startCursor === undefined
+            ? connection.pageInfo.startCursor
+            : state.pageInfo.startCursor;
 
-      let pageEndCursor =
-        state?.pageInfo?.endCursor === undefined
-          ? connection.pageInfo.endCursor
-          : state.pageInfo.endCursor;
+        let pageEndCursor =
+          state?.pageInfo?.endCursor === undefined
+            ? connection.pageInfo.endCursor
+            : state.pageInfo.endCursor;
 
-      let previousPageExists =
-        state?.pageInfo?.hasPreviousPage === undefined
-          ? connection.pageInfo.hasPreviousPage
-          : state.pageInfo.hasPreviousPage;
+        let previousPageExists =
+          state?.pageInfo?.hasPreviousPage === undefined
+            ? connection.pageInfo.hasPreviousPage
+            : state.pageInfo.hasPreviousPage;
 
-      let nextPageExists =
-        state?.pageInfo?.hasNextPage === undefined
-          ? connection.pageInfo.hasNextPage
-          : state.pageInfo.hasNextPage;
+        let nextPageExists =
+          state?.pageInfo?.hasNextPage === undefined
+            ? connection.pageInfo.hasNextPage
+            : state.pageInfo.hasNextPage;
 
-      if (state?.nodes) {
         if (isPrevious) {
           pageStartCursor = connection.pageInfo.startCursor;
           previousPageExists = connection.pageInfo.hasPreviousPage;
@@ -216,16 +280,24 @@ export function usePagination<NodesType>(
           pageEndCursor = connection.pageInfo.endCursor;
           nextPageExists = connection.pageInfo.hasNextPage;
         }
-      }
 
+        setCurrentPageInfo({
+          startCursor: pageStartCursor,
+          endCursor: pageEndCursor,
+          hasPreviousPage: previousPageExists,
+          hasNextPage: nextPageExists,
+        });
+      }
+    } else {
+      setNodes(flattenConnection(connection));
       setCurrentPageInfo({
-        startCursor: pageStartCursor,
-        endCursor: pageEndCursor,
-        hasPreviousPage: previousPageExists,
-        hasNextPage: nextPageExists,
+        startCursor: connection.pageInfo.startCursor,
+        endCursor: connection.pageInfo.endCursor,
+        hasPreviousPage: connection.pageInfo.hasPreviousPage,
+        hasNextPage: connection.pageInfo.hasNextPage,
       });
     }
-  }, [state, connection]);
+  }, [state, connection, isPrevious, search, navigate, pathname]);
 
   const previousPageUrl = useMemo(() => {
     const params = new URLSearchParams(search);
@@ -256,7 +328,7 @@ export function getPaginationVariables(
   request: Request,
   options: {pageBy: number} = {pageBy: 20},
 ) {
-  if (!(request instanceof Request)) {
+  if (typeof request?.url === 'undefined') {
     throw new Error(
       'getPaginationVariables must be called with the Request object passed to your loader function',
     );