+++ templates/hello-world/server.ts
@@ -53,7 +53,7 @@ export default {
       const handleRequest = createRequestHandler({
         build: remixBuild,
         mode: process.env.NODE_ENV,
-        getLoadContext: () => ({session, storefront, env}),
+        getLoadContext: () => ({session, storefront, env, waitUntil}),
       });
 
       const response = await handleRequest(request);