+++ templates/skeleton/app/routes/collections.$handle.tsx
@@ -174,8 +174,8 @@ const COLLECTION_QUERY = `#graphql
         pageInfo {
           hasPreviousPage
           hasNextPage
-          hasNextPage
           endCursor
+          startCursor
         }
       }
     }